feat: footer first adaptive layout
