+++ src/components/Footer/Footer.module.scss
@@ -1,0 +1,169 @@
+@import '../../utils/scss-variables.scss';
+
+.bg {
+  width: 100%;
+  display: flex;
+  padding: 70px 0;
+  justify-content: center;
+  align-items: center;
+  background-color: $main-black;
+  color: $main-grey-light;
+
+  @media screen and (max-width: 900px) {
+    padding: 60px 0;
+  }
+  @media screen and (max-width: 670px) {
+    padding: 33px 16px 68px 16px;
+  }
+}
+
+.container {
+  width: fit-content;
+  display: grid;
+  grid-template-columns: repeat(3, min-content);
+  grid-template-rows: 1fr;
+  grid-auto-columns: max-content;
+  column-gap: 204px;
+  height: auto;
+
+  @media screen and (max-width: 1280px) {
+    column-gap: 80px;
+  }
+  @media screen and (max-width: 670px) {
+    grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+    grid-template-rows: auto auto;
+    row-gap: 32px;
+    column-gap: 50px;
+    width: 100%;
+  }
+  @media screen and (max-width: 374px) {
+    column-gap: 10px;
+  }
+  @media screen and (max-width: 320px) {
+    column-gap: 5px;
+  }
+}
+
+.logo {
+  grid-column: 1;
+  font-size: 20px;
+  width: min-content;
+  margin: 0;
+  padding: 0;
+  font-family: 'Gilroy';
+  font-style: normal;
+  font-weight: 700;
+  font-size: 20px;
+  line-height: 25px;
+}
+
+.nav {
+  grid-column: 2;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  width: 471px;
+  gap: 0;
+
+  @media screen and (max-width: 1280px) {
+    width: 380px;
+  }
+  @media screen and (max-width: 900px) {
+    flex-direction: column;
+    width: min-content;
+    align-items: start;
+    gap: 32px;
+  }
+  @media screen and (max-width: 670px) {
+    grid-column: 1;
+    grid-row: 2;
+    gap: 21px;
+  }
+
+  .nav_item {
+    text-decoration: none;
+    color: $main-grey-light;
+    font-family: 'Arimo';
+    font-style: normal;
+    font-weight: 400;
+    font-size: 14px;
+    line-height: 16px;
+  }
+  .nav_item:hover {
+    color: $accent;
+  }
+  .nav_item:active {
+    color: $accent;
+    text-decoration: underline;
+    text-underline-offset: 4px;
+  }
+}
+
+.sm_block {
+  grid-column: 3;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 144px;
+  justify-self: end;
+
+  @media screen and (max-width: 900px) {
+    align-items: start;
+  }
+
+  @media screen and (max-width: 670px) {
+    grid-column: 2;
+    grid-row: 1;
+    width: 121px;
+  }
+
+  @media screen and (max-width: 320px) {
+    width: 90px;
+  }
+
+  .sm_vk {
+    display: block;
+    width: 30px;
+    height: 25px;
+    background-image: url(../../images/vk.svg);
+    background-repeat: no-repeat;
+    background-position: center;
+    cursor: pointer;
+    
+    @media screen and (max-width: 670px) {
+      width: 20px;
+      height: 25px;
+    }
+  }
+
+  .sm_inst {
+    display: block;
+    width: 25px;
+    height: 25px;
+    background-image: url(../../images/inst.svg);
+    background-repeat: no-repeat;
+    background-position: center;
+    cursor: pointer;
+
+    @media screen and (max-width: 670px) {
+      width: 17px;
+      height: 25px;
+    }
+  }
+
+  .sm_tg {
+    display: block;
+    width: 28px;
+    height: 25px;
+    background-image: url(../../images/tg.svg);
+    background-repeat: no-repeat;
+    background-position: center;
+    cursor: pointer;
+
+    @media screen and (max-width: 670px) {
+      width: 18px;
+      height: 25px;
+    }
+  }
+}
