--- src/pages/Header/Header.css
@@ -1,175 +1,0 @@
-.header {
-  box-sizing: border-box;
-  width: 100%;
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  padding: 24px 17px;
-  background: #ffffff;
-  box-shadow: 0px 0.996045px 3.98418px rgba(51, 51, 51, 0.15);
-}
-
-.header__container {
-  max-width: 1175px;
-  margin: 0 auto;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.header__item {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  padding: 0px;
-  gap: 32px;
-}
-
-.header__image {
-  font-family: "Arimo";
-  font-style: normal;
-  font-weight: 700;
-  font-size: 24px;
-  line-height: 28px;
-  color: #343843;
-  text-decoration: none;
-}
-
-.header__locations {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  padding: 0px;
-  gap: 4px;
-}
-
-.header__location {
-  background-image: url(../../images/location.svg);
-  width: 36px;
-  height: 36px;
-}
-
-.header__city {
-  font-family: "Gilroy";
-  font-style: normal;
-  font-weight: 400;
-  font-size: 18px;
-  line-height: 140%;
-  color: #343843;
-}
-
-.header__navigations {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  padding: 0px;
-  gap: 40px;
-}
-
-.header__navigation {
-  font-family: "Gilroy";
-  font-style: normal;
-  font-weight: 300;
-  font-size: 14px;
-  line-height: 17px;
-  text-align: center;
-  color: #343843;
-  margin: 0;
-}
-.header__navigation:hover {
-  font-family: "Gilroy";
-  font-style: normal;
-  font-weight: 700;
-  font-size: 14px;
-  line-height: 17px;
-  text-align: center;
-  color: #2fafff;
-}
-.header__navigation_active {
-  font-family: "Gilroy";
-  font-style: normal;
-  font-weight: 700;
-  font-size: 14px;
-  line-height: 17px;
-  text-align: center;
-  -webkit-text-decoration-line: underline;
-  text-decoration-line: underline;
-  color: #2fafff;
-}
-.header__additional {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  padding: 0px;
-  gap: 24px;
-}
-.header__loop {
-  background-image: url(../../images/search.svg);
-  width: 24px;
-  height: 24px;
-}
-.header__languages {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  padding: 0px 0px 0px 4px;
-  gap: 8px;
-  width: 80px;
-  height: 33px;
-  border: none;
-  font-family: "Arimo";
-  font-style: normal;
-  font-weight: 400;
-  font-size: 14px;
-  line-height: 120.02%;
-  color: #343843;
-}
-.header__language {
-  font-family: "Arimo";
-  font-style: normal;
-  font-weight: 400;
-  font-size: 14px;
-  line-height: 120.02%;
-  color: #343843;
-  margin: 0;
-  position: relative;
-}
-.header__flag {
-  display: block;
-  width: 24px;
-  height: 24px;
-}
-.header__flag_ru {
-  background-image: url(../../images/lang_ru.svg);
-  width: 24px;
-  height: 24px;
-}
-.header__language-text {
-  font-family: "Arimo";
-  font-style: normal;
-  font-weight: 400;
-  font-size: 14px;
-  line-height: 120.02%;
-  color: #343843;
-  margin: 0;
-}
-.header__button {
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  /* padding: 12px 0; */
-  gap: 10px;
-  width: 185px;
-  height: 46px;
-  border: 2px solid #2fafff;
-  border-radius: 8px;
-  font-family: "Gilroy";
-  font-style: normal;
-  font-weight: 400;
-  font-size: 18px;
-  line-height: 22px;
-  color: #343843;
-}
